verifyCookies function to GlobalNotification component
Wanted to make it an action in the user store, but can't figure out how to make actions in stores call actions from other stores.
