+++ frontend/src/components/globalNotification/index.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react'
 import { inject, observer } from 'mobx-react'
 import { onSnapshot } from 'mobx-state-tree'
 import { PopUp } from '../../components'
+import {popUpData} from "../popUp/popUpData";
 
 @inject('store') @observer
 class GlobalNotification extends Component {
@@ -9,6 +10,7 @@ class GlobalNotification extends Component {
 		super(props)
 
 		this.initNotificationListener()
+		this.verifyCookies()
 
 		this.state = {
 			isNotificationVisible: this.props.store.notification.visible
@@ -22,6 +24,26 @@ class GlobalNotification extends Component {
 					isNotificationVisible: notification.visible
 				})
 			}
+		})
+	}
+
+	verifyCookies = () => {
+		const cookiesAcceptedAt = localStorage.getItem('cookiesAcceptedAt')
+
+		const SEVEN_DAYS = 60 * 60 * 1000 * 24 * 7
+
+		if (cookiesAcceptedAt) {
+			if (new Date() - SEVEN_DAYS >= Date.parse(cookiesAcceptedAt)) {
+				localStorage.removeItem('cookiesAcceptedAt')
+			} else {
+				return
+			}
+		}
+
+		this.store.notification.setContent(popUpData.messages.cookies)
+
+		this.store.notification.setCustomClose(() => {
+			localStorage.setItem('cookiesAcceptedAt', new Date().toString())
 		})
 	}
